Fixes pointed out by clang-tidy (#310)
* Rearrange header includes so clang-tidy is happy.

The ament checkers are also happy with this arrangement.


* Get rid of unnecessary void arguments on methods.


* Properly mark methods as override where appropriate.


* Initialize members and stack variables to zero before use.


* Mark function implementations in header files 'inline'.


* Use more efficient 'empty' method instead of empty string.


* Get rid of unnecessary else statements after a continue/return.


* Get rid of unnecessary == false use.


* Make sure to add a virtual destructor to DataContainerBase.


* Use string != comparison instead of 'compare' method.


* Rename PointcloudXYZIR parameter name to match implementation.


* Get rid of usage of typedef.

Also use C++-style reinterpret_cast instead of C-style cast.


* Be more explicit about using floats.


* Do an explicit lround to do int->float conversion.


* Fix the intensity calculation.

There are a few problems.  First is that it was doing an
integer divide, which seems like it would be always 0.  Second,
the number that was being used was the raw integer, not the
distance.  Finally, the order of operations meant that it wasn't
actually implementing the algorithm shown in the VLP documentation.
This commit fixes all of those problems.


* Remove the azimuth from addPoint.

None of the implementations use it.


* Make sure to reset the cloud data to 0 before reusing.

We were seeing artifacts in the organize_cloud case that this removes.


* Use underscores on pointcloud member variables.


* Using std::lround means we don't need to add 0.5 to the result.


* Slightly rearrange and simplify range checks.


* Add in example launch files for VLP32C.

+++ velodyne_pointcloud/src/lib/calibration.cpp
@@ -140,7 +140,7 @@ void operator>>(const YAML::Node & node, std::pair<int, LaserCorrection> & corre
 
 #endif
 
-  correction.second.max_intensity = ::floor(max_intensity_float);
+  correction.second.max_intensity = ::floorf(max_intensity_float);
 
   float min_intensity_float = 0.0;
 
@@ -158,7 +158,7 @@ void operator>>(const YAML::Node & node, std::pair<int, LaserCorrection> & corre
 
 #endif
 
-  correction.second.min_intensity = ::floor(min_intensity_float);
+  correction.second.min_intensity = ::floorf(min_intensity_float);
 
   node[FOCAL_DISTANCE] >> correction.second.focal_distance;
   node[FOCAL_SLOPE] >> correction.second.focal_slope;
